# networking: speed up refreshing