--- a/selfdrive/ui/qt/offroad/wifiManager.cc
+++ b/selfdrive/ui/qt/offroad/wifiManager.cc
@@ -20,14 +20,6 @@ T get_response(QDBusMessage response) {
   }
 }
 
-bool compare_by_strength(const Network &a, const Network &b) {
-  if (a.connected == ConnectedType::CONNECTED) return true;
-  if (b.connected == ConnectedType::CONNECTED) return false;
-  if (a.connected == ConnectedType::CONNECTING) return true;
-  if (b.connected == ConnectedType::CONNECTING) return false;
-  return a.strength > b.strength;
-}
-
 WifiManager::WifiManager(QWidget* parent) : QWidget(parent) {
   qDBusRegisterMetaType<Connection>();
   qDBusRegisterMetaType<IpConfig>();
@@ -74,40 +66,49 @@ void WifiManager::setup() {
   requestScan();
 }
 
+void WifiManager::addNetwork(const QDBusObjectPath &path) {
+  const QByteArray &ssid = get_property(path.path(), "Ssid");
+  if (ssid.isEmpty()) {
+    return;
+  }
+  const unsigned int strength = get_ap_strength(path.path());
+  const SecurityType &security = getSecurityType(path.path());
+  ConnectedType ctype;
+  if (path.path() != activeAp) {
+    ctype = ConnectedType::DISCONNECTED;
+  } else {
+    if (ssid == connecting_to_network) {
+      ctype = ConnectedType::CONNECTING;
+    } else {
+      ctype = ConnectedType::CONNECTED;
+    }
+  }
+  Network network = {path.path(), ssid, strength, ctype, security, isKnownConnection(ssid)};
+  if (!seenNetworks.contains(ssid)) {
+    mutex.lock();
+    seenNetworks[ssid] = network;
+    mutex.unlock();
+  }
+}
+
 void WifiManager::refreshNetworks() {
   if (adapter.isEmpty()) {
     return;
   }
-  seen_networks.clear();
-  seen_ssids.clear();
+  seenNetworks.clear();
   ipv4_address = get_ipv4_address();
 
   QDBusInterface nm(NM_DBUS_SERVICE, adapter, NM_DBUS_INTERFACE_DEVICE_WIRELESS, bus);
   nm.setTimeout(DBUS_TIMEOUT);
 
+  std::vector<std::thread> threads;
   const QDBusReply<QList<QDBusObjectPath>> &response = nm.call("GetAllAccessPoints");
   for (const QDBusObjectPath &path : response.value()) {
-    QByteArray ssid = get_property(path.path(), "Ssid");
-    if (ssid.isEmpty() || seen_ssids.contains(ssid)) {
-      continue;
-    }
-    unsigned int strength = get_ap_strength(path.path());
-    SecurityType security = getSecurityType(path.path());
-    ConnectedType ctype;
-    if (path.path() != activeAp) {
-      ctype = ConnectedType::DISCONNECTED;
-    } else {
-      if (ssid == connecting_to_network) {
-        ctype = ConnectedType::CONNECTING;
-      } else {
-        ctype = ConnectedType::CONNECTED;
-      }
-    }
-    Network network = {path.path(), ssid, strength, ctype, security};
-    seen_ssids.push_back(ssid);
-    seen_networks.push_back(network);
+    threads.push_back(std::thread(&WifiManager::addNetwork, this, path));
   }
-  std::sort(seen_networks.begin(), seen_networks.end(), compare_by_strength);
+  for (auto &th : threads) {
+    th.join();
+  }
 }
 
 QString WifiManager::get_ipv4_address() {