Pools threads to get attributes of each AP.

For 16-20 networks, previous times per network were: 10.23 ms, 10.9 ms, 11.31 ms, 10.97 ms, 11.3219 ms, 12.01 ms, 12.31 ms (mean: 11.29 ms)
For 11-17 networks now: 3.27 ms, 3.787 ms, 3.12 ms, 2.65 ms, 3.05 ms, 2.72 ms, 2.48 ms (mean: 3.011 ms)

~3.75x speed increase

Current times:
```
selfdrive/ui/qt/offroad/networking.cc: Refreshing took 73.651 ms for 25 networks - 2.94604 ms/network
selfdrive/ui/qt/offroad/networking.cc: Drawing took 41.6182 ms for 25 networks - 1.66473 ms/network
```